Improved error reporting in asm.lua
Specifically, I went through and made all asserts and errors
return the same consistant 'return false, "error message"' form
(which was very tedious), cleaned up an artefact in scrub that
was preventing proper line numbers, and made the stand-alone
runnable of asm report the given error when assembly fails.

All in all, this should be a notable improvement.