added LOAD and STOR opcodes
LOAD and STOR are implemented as free-register opcodes, accepting
either two registers, or one register and a number. the load a value
from memory into a register, and store a value from a register into
memory respectively.

they should be used in preference to the assignment forms of MOV
(which will be deprecated)